+++ examples/spot/margin/get_margin_oco_order.py
@@ -10,18 +10,12 @@ config_logging(logging, logging.DEBUG)
 key = ""
 secret = ""
 
-client = Client(key, secret, base_url="https://testnet.binance.vision")
+client = Client(key, secret)
 
 try:
-    # get oco orders with default parameters
-    response = client.get_oco_orders()
-    logging.info(response)
-
-    # provide parameters
-    response = client.get_oco_orders(fromId=1, limit=10)
-    logging.info(response)
-
-    response = client.get_oco_orders(startTime=1588160817647, limit=10)
+    response = client.get_margin_oco_order(
+        origClientOrderId="ARzZ9I00CPM8i3Nh", isIsolated="TRUE", symbol="BNBUSDT"
+    )
     logging.info(response)
 except ClientError as error:
     logging.error(