+++ tests/spot/margin/test_new_margin_oco_order.py
@@ -1,5 +1,4 @@
 import responses
-
 from tests.util import random_str
 from urllib.parse import urlencode
 from tests.util import mock_http_response
@@ -24,22 +23,22 @@ params = {
 }
 
 
-@mock_http_response(responses.POST, "/api/v3/order/oco", mock_exception, 400)
-def test_post_an_oco_order_without_param():
-    """Tests the API endpoint to post a new oco order without parameters"""
+@mock_http_response(responses.POST, "/sapi/v1/margin/order/oco", mock_exception, 400)
+def test_post_an_margin_oco_order_without_params():
+    """Tests the API endpoint to post a new margin oco order without parameters"""
 
     client = Client(key, secret)
-    client.new_oco_order.when.called_with("", "", "", "", "").should.throw(
+    client.new_margin_oco_order.when.called_with("", "", "", "", "").should.throw(
         ParameterRequiredError
     )
 
 
 @mock_http_response(
-    responses.POST, "/api/v3/order/oco\\?" + urlencode(params), mock_item, 200
+    responses.POST, "/sapi/v1/margin/order/oco\\?" + urlencode(params), mock_item, 200
 )
-def test_post_an_oco_order():
-    """Tests the API endpoint to post a new oco order"""
+def test_post_an_margin_oco_order():
+    """Tests the API endpoint to post a new margin oco order"""
 
     client = Client(key, secret)
-    response = client.new_oco_order(**params)
+    response = client.new_margin_oco_order(**params)
     response.should.equal(mock_item)